requirements.txt
openpyxl==3.1.5
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
google-generativeai==0.8.3
python-dotenv==1.0.1
streamlit==1.43.2
altair
